feat: implementar módulo Alimento completo con CRUD y validaciones
Módulo Rutina completo:
- 2 DTOs
- Alimento con MapStruct
- IAlimentoRepository con 8 métodos de consulta
- IAlimentoService con 15 métodos
- AlimentoService con validaciones, transacciones y logs
- AlimentoController con 11 endpoints REST

Mejoras en controllers existentes:
- Usuario: Validación de parámetros entre username y email
- Rutina: Validación de parámetros en nivel y nombre
- Ejercicio: Validación de parámmetros en grupo muscular, dificultad y nombre
- Todos los controllers: Validación consistente de listas vacías

Cambio en base de datos en la entidad Alimento:
- Cambio de columna proteínas de INT a DECIMAL(5,2)

Corrección de configuración:
- Logback: Eliminado triggeringPolicy que causaba NullPointerException
- Logback: Configuración simplificada con TimeBasedRollingPolicy
- Logs: Rolling diario sin límite de tamaño por archivo

Todo comprobado y funcionando correctamente.

+++ api/gymprofit-api/src/main/java/com/gymprofit/api/entity/Alimento.java
@@ -30,7 +30,7 @@ public class Alimento {
     private Integer calorias;
 
     @Column(precision = 5, scale = 2)
-    private Integer proteinas;
+    private BigDecimal proteinas;
 
     @Column(precision = 5, scale = 2)
     private BigDecimal carbohidratos;
